+++ test/webrtc-libc-wasm/wasm/src/lib.rs
@@ -21,6 +21,7 @@ use flmodules::{
 };
 
 const URL: &str = "ws://127.0.0.1:8765";
+const MODULE_NAME: &str = "LIBC_WASM";
 
 #[derive(Debug, Error)]
 enum StartError {
@@ -72,7 +73,11 @@ async fn run_app() -> anyhow::Result<()> {
                             Destination::NoTap,
                             NetworkIn::MessageToNode(
                                 other.get(0).unwrap().get_id(),
-                                "Hello from Rust wasm".to_string(),
+                                NetworkWrapper::wrap_yaml(
+                                    MODULE_NAME,
+                                    &"Hello from Rust wasm".to_string(),
+                                )
+                                .expect("Creating NetworkWrapper"),
                             )
                             .into(),
                         )?;