+++ CHANGELOG.md
@@ -24,6 +24,7 @@ See [WIP.md](./WIP.md) for current work and bugs to fix
 - adding TimerBroker to network_start
 - flarch::Browser got macros for handling forward messages
 - WebRTCConnInput, WebRTCConnOutput to WebRTCConnIn, WebRTCConnOut
+- Adding NetworkMessage to NetworkOut
 
 ### Fixed